Refactoring inspectors' structure
