+++ sdgx/data_models/inspectors/base.py
@@ -1,8 +1,13 @@
 from __future__ import annotations
 
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import pandas as pd
+
+if TYPE_CHECKING:
+    from sdgx.data_models.metadata import Metadata
+
+from sdgx.data_models.relationship import Relationship
 
 
 class Inspector:
@@ -28,3 +33,28 @@ class Inspector:
 
     def inspect(self, *args, **kwargs) -> dict[str, Any]:
         """Inspect raw data and generate metadata."""
+
+
+class RelationshipInspector(Inspector):
+    """
+    Empty RelationshipInspector for inheritence
+
+    Subclass should implement `_build_relationship` and `fit`
+    """
+
+    def _build_relationship(self) -> list[Relationship]:
+        return []
+
+    def fit(
+        self,
+        raw_data: pd.DataFrame,
+        name: str | None = None,
+        metadata: "Metadata" | None = None,
+        *args,
+        **kwargs,
+    ):
+        pass
+
+    def inspect(self, *args, **kwargs) -> dict[str, Any]:
+        """Inspect raw data and generate metadata."""
+        return {"relationships": self._build_relationship()}
