+++ sdgx/data_models/metadata.py
@@ -11,8 +11,8 @@ import pandas as pd
 from pydantic import BaseModel
 
 from sdgx.data_loader import DataLoader
+from sdgx.data_models.inspectors.base import RelationshipInspector
 from sdgx.data_models.inspectors.manager import InspectorManager
-from sdgx.data_models.inspectors.relationship import RelationshipInspector
 from sdgx.exceptions import MetadataInitError, MetadataInvalidError
 from sdgx.utils import logger
